% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecs_all.R
\name{oecs_all}
\alias{oecs_all}
\title{Set up an new screening.}
\usage{
oecs_all(file = "oecs.xlsx")
}
\arguments{
\item{file}{Path to the .xlsx-file. If the file does not exist, it will be created.}
}
\value{
Number of the Survey-ID.
}
\description{
Set up an new screening.
}
\examples{
add(1, 1)
}
