% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecs_evaluate.R
\name{oecs_evaluate}
\alias{oecs_evaluate}
\title{Set up an new screening.}
\arguments{
\item{adress}{dataframe with adresses of the childs}

\item{survey_id}{vector with strings, choosing the scales ("sprachstand","umwelt","who")}

\item{save}{boolean saving files}

\item{path_answers}{string path and filename of participants}

\item{path_pending}{string path and filename of important settings (survey id, salt of hash)}

\item{path_not_assignable}{string path and filename of important settings (survey id, salt of hash)}
}
\value{
List with 3 dataframes (answers, pending, not assignable)
}
\description{
Set up an new screening.
}
\examples{
add(1, 1)
}
