% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecs_init.R
\name{oecs_init}
\alias{oecs_init}
\title{Set up an new screening.}
\arguments{
\item{adress}{dataframe with adresses of the childs}

\item{scales}{vector with strings, choosing the scales ("sprachstand","umwelt","who")}

\item{activate}{boolean for activateing the survey}

\item{save}{boolean saving files}

\item{path_participants}{string path and filename of participants}

\item{path_settings}{string path and filename of important settings (survey id, salt of hash)}

\item{survey_name}{string name survey}
}
\value{
Number of the Survey-ID.
}
\description{
Set up an new screening.
}
\examples{
add(1, 1)
}
