% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecs_evaluate.R, R/oecs_letter.R
\name{oecs_evaluate}
\alias{oecs_evaluate}
\title{Set up an new screening.}
\arguments{
\item{survey_id}{integer number of survey}

\item{save}{boolean saving files}

\item{path_answers}{string path and filename of answers}

\item{path_pending}{string path and filename of pending}

\item{path_not_assignable}{string path and filename of not assignable}

\item{adress}{dataframe with adresses of the childs}

\item{survey_url}{vector with strings, choosing the scales ("sprachstand","umwelt","who")}

\item{path_letter}{string path (no filename) to save the files for the letters}
}
\value{
List with 3 dataframes (answers, pending, not assignable)

List with 3 dataframes (answers, pending, not assignable)
}
\description{
Set up an new screening.

Set up an new screening.
}
\examples{
add(1, 1)
add(1, 1)
}
